# cabinet-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CabinetPro API",
        "description": "API REST de gestion de cabinet d'avocat — CabinetPro.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/api/v1/factures": {
            "get": {
                "tags": [
                    "Facturation"
                ],
                "summary": "Liste des factures",
                "operationId": "9f77f3471a110c9fd8c1f9c6d82e9e25",
                "parameters": [
                    {
                        "name": "statut",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "client_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagination des factures"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Facturation"
                ],
                "summary": "Créer une facture avec ses lignes",
                "operationId": "6fc95a208a56c44157c7a214ee6a2439",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Facture"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Facture créée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/factures/stats": {
            "get": {
                "tags": [
                    "Facturation"
                ],
                "summary": "Statistiques de facturation",
                "operationId": "ee1e7e27648c44dd5022d597a2f0dd13",
                "responses": {
                    "200": {
                        "description": "Statistiques"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/factures/{facture}": {
            "get": {
                "tags": [
                    "Facturation"
                ],
                "summary": "Détail d'une facture",
                "operationId": "c4ad48cad56c060b600a053ec88f441b",
                "parameters": [
                    {
                        "name": "facture",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Facture"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Facturation"
                ],
                "summary": "Modifier une facture",
                "operationId": "c092a3bdba1fe68a6112424f58fe3743",
                "parameters": [
                    {
                        "name": "facture",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Facture"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Facture modifiée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Facturation"
                ],
                "summary": "Supprimer une facture",
                "operationId": "3384c684f5ddac9968b2a8fa5218f6a9",
                "parameters": [
                    {
                        "name": "facture",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Facture supprimée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/factures/{facture}/paiements": {
            "post": {
                "tags": [
                    "Facturation"
                ],
                "summary": "Enregistrer un paiement",
                "operationId": "781a9713e32354f5d05aa9d7b24e92d3",
                "parameters": [
                    {
                        "name": "facture",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Paiement"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paiement enregistré"
                    },
                    "422": {
                        "description": "Montant supérieur au solde"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/factures/{facture}/pdf": {
            "get": {
                "tags": [
                    "Facturation"
                ],
                "summary": "Télécharger le PDF d'une facture",
                "operationId": "03dfd6a975911077a654e635a97cd525",
                "parameters": [
                    {
                        "name": "facture",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier PDF"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Facture": {
                "required": [
                    "client_id",
                    "date_facture",
                    "date_echeance",
                    "lignes"
                ],
                "properties": {
                    "client_id": {
                        "type": "integer"
                    },
                    "dossier_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "date_facture": {
                        "type": "string",
                        "format": "date"
                    },
                    "date_echeance": {
                        "type": "string",
                        "format": "date"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "brouillon",
                            "en_attente",
                            "payee",
                            "annulee"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "lignes": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "designation": {
                                    "type": "string"
                                },
                                "quantite": {
                                    "type": "integer"
                                },
                                "prix_unitaire": {
                                    "description": "En Ariary (entier)",
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Paiement": {
                "required": [
                    "montant",
                    "mode",
                    "date_paiement"
                ],
                "properties": {
                    "montant": {
                        "description": "En Ariary (entier)",
                        "type": "integer"
                    },
                    "mode": {
                        "type": "string",
                        "enum": [
                            "especes",
                            "virement",
                            "mobile_money",
                            "cheque"
                        ]
                    },
                    "date_paiement": {
                        "type": "string",
                        "format": "date"
                    },
                    "reference": {
                        "type": "string",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}